test(blinc_cn): reorganize responsive tests and cover scroll defaults
